Move PlayerInputManager logic inside RulesetInputManager
